Improve sparse file handling.
Clients now send the index inside the collected file buffers. Clients
also keep the file size and the stored size separate and report
both. This way we can track the uploaded bytes accurates as the total
bytes we expect, but have the file size also explicitely shown.

+++ services/server_artifacts/server_uploader.go
@@ -52,8 +52,8 @@ func (self *ServerUploader) Upload(
 		)
 		self.collection_context.Modify(func(context *flows_proto.ArtifactCollectorContext) {
 			context.TotalUploadedFiles++
-			context.TotalUploadedBytes += result.Size
-			context.TotalExpectedUploadedBytes += result.Size
+			context.TotalUploadedBytes += uint64(result.Size)
+			context.TotalExpectedUploadedBytes += uint64(result.Size)
 		})
 
 		err = self.collection_context.Save()
